requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1
scipy>=1.7.0

# Deep learning
torch>=1.9.0
torchvision>=0.11.0
torchaudio>=0.10.0

# Data acquisition
yfinance>=0.1.63
kaggle>=1.5.12

# Feature engineering
pandas-ta>=0.3.14b0
ta>=0.10.0
statsmodels>=0.13.0

# Hyperparameter optimization
optuna>=2.10.0

# Visualization
plotly>=5.1.0
dash>=2.0.0
streamlit>=1.0.0

# API and deployment
flask>=2.0.0
fastapi>=0.75.0
uvicorn>=0.17.0

# Configuration and utilities
pyyaml>=6.0
joblib>=1.1.0
tqdm>=4.62.0
python-dotenv>=0.19.0

# Testing
pytest>=6.2.4
pytest-cov>=2.12.0
black>=21.6b0
flake8>=3.9.2
